move theme contants/types to types.ts
