+++ src/options/types.ts
@@ -15,23 +15,6 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import { getSettingsState } from "@/store/settingsStorage";
-import { getThemeLogo } from "@/utils/themeUtils";
-import activateBrowserActionIcon from "@/background/activateBrowserActionIcon";
-import { DEFAULT_THEME } from "@/options/types";
-
-async function setToolbarIcon(): Promise<void> {
-  const { theme } = await getSettingsState();
-
-  if (theme === DEFAULT_THEME) {
-    activateBrowserActionIcon();
-    return;
-  }
-
-  const themeLogo = getThemeLogo(theme);
-  (chrome.browserAction ?? chrome.action).setIcon({ path: themeLogo.small });
-}
-
-export default function initPartnerTheme() {
-  void setToolbarIcon();
-}
+export const DEFAULT_THEME = "default";
+export const THEMES = [DEFAULT_THEME, "automation-anywhere"] as const;
+export type Theme = typeof THEMES[number];
